app: background-isolate FFI worker, persistent app-support data root, full journey UI, no-op sweep; CI fmt/clippy/analyze/test gates
- harbor_native: HarborCoreWorker (long-lived isolate owning the native
  handle; JSON requests across a command port; harbor_core_open_ex)
- harbor_app: async HarborService over the worker (unique random run
  ids, op progress polling with kind-tracked snapshots, cancel,
  knowledge sources/remove, path-based model install, HF file listing);
  main.dart bootstraps the persistent data root via path_provider,
  closes the core on detach/dispose, honest boot/degraded states
- surfaces: Ask runs the full grounded journey (run + generation op +
  citations + token progress + cancel); Knowledge manages real sources
  (add files/paste text with ingest progress, remove, identity);
  Models acquires HF packages with real byte progress + local GGUF
  import; Home chips prefill the composer, attach routes files into
  the index, ModelDock navigates; Agents shows an honest disabled
  state; Activity replays async; Settings shows the durable device id
- android: MainActivity AndroidKeyStore adapter (non-exportable AES-GCM
  wrap of the device root; plaintext only in memory per launch)
- l10n: EN/AR parity for all new strings (gen-l10n regenerated)
- tests: shell/a11y suites moved to the async API; knowledge replace
  + durable-run replay coverage added; dart format + flutter analyze
  clean in all four Dart packages
- ci: rustfmt, clippy -D warnings, dart format, flutter analyze, and
  live-core flutter test gates added

+++ core/harbor_render/src/preview.rs
@@ -100,7 +100,7 @@ impl WorkbookPreview {
             });
         }
         cells.sort_by_key(|c| (c.row, c.col));
-        let chart_count = WorkbookDoc::count_charts_in_bytes(bytes)? as usize;
+        let chart_count = WorkbookDoc::count_charts_in_bytes(bytes)?;
         Ok(WorkbookPreview {
             kind: "workbook".into(),
             sheet: first,
@@ -153,7 +153,11 @@ mod tests {
         let preview = WorkbookPreview::from_xlsx(&bytes).unwrap();
         assert_eq!(preview.kind, "workbook");
         assert_eq!(preview.sheet, "Sheet1");
-        let a2 = preview.cells.iter().find(|c| c.row == 2 && c.col == 1).unwrap();
+        let a2 = preview
+            .cells
+            .iter()
+            .find(|c| c.row == 2 && c.col == 1)
+            .unwrap();
         assert!(a2.formula.as_deref().unwrap_or_default().contains("A1*3"));
         assert_eq!(a2.value.as_deref(), Some("30"));
         let json = serde_json::to_string(&preview).unwrap();
@@ -169,7 +173,7 @@ mod tests {
                 bullets: vec!["Q1 3600".into()],
                 notes: None,
                 chart: None,
-            image: None,
+                image: None,
             }],
         };
         let bytes = deck.to_pptx_bytes().unwrap();
